// src/_services/tg-bot/__generated__/SetChatStickerSet.ts
/* eslint-disable */
/* tslint:disable */
// @ts-nocheck
/*
 * ---------------------------------------------------------------
 * ## THIS FILE WAS GENERATED VIA SWAGGER-TYPESCRIPT-API        ##
 * ##                                                           ##
 * ## AUTHOR: acacode                                           ##
 * ## SOURCE: https://github.com/acacode/swagger-typescript-api ##
 * ---------------------------------------------------------------
 */

import {
  SetChatStickerSetCreateData,
  SetChatStickerSetCreateError,
  SetChatStickerSetCreatePayload,
} from "./data-contracts";
import { ContentType, HttpClient, RequestParams } from "./http-client";

export class SetChatStickerSet<SecurityDataType = unknown> {
  http: HttpClient<SecurityDataType>;

  constructor(http: HttpClient<SecurityDataType>) {
    this.http = http;
  }

  /**
   * @description Use this method to set a new group sticker set for a supergroup. The bot must be an administrator in the chat for this to work and must have the appropriate admin rights. Use the field *can\_set\_sticker\_set* optionally returned in [getChat](https://core.telegram.org/bots/api/#getchat) requests to check if the bot can use this method. Returns *True* on success.
   *
   * @name SetChatStickerSetCreate
   * @request POST:/setChatStickerSet
   */
  setChatStickerSetCreate = (
    data: SetChatStickerSetCreatePayload,
    params: RequestParams = {},
  ) =>
    this.http.request<
      SetChatStickerSetCreateData,
      SetChatStickerSetCreateError
    >({
      path: `/setChatStickerSet`,
      method: "POST",
      body: data,
      type: ContentType.Json,
      format: "json",
      ...params,
    });
}
